[fix]型宣言ファイルの配置修正
