+++ src/types/api/gameDetailsType.ts
@@ -1,18 +1,4 @@
-export type steamGameCategoryType = {
-  id: number,
-  description: string
-}
-
-export type steamGameGenreType = {
-  id: string,
-  description: string
-}
-
-export type steamGamePlatformType = {
-  windows: boolean,
-  mac: boolean,
-  linux: boolean
-}
+import { steamGamePlatformType } from "./steamDataType"
 
 export type gameDetailType = {
   twitchGameId: string,
