+++ src/types/api/steamDataType.ts
@@ -1,0 +1,15 @@
+export type steamGameCategoryType = {
+  id: number,
+  description: string
+}
+
+export type steamGameGenreType = {
+  id: string,
+  description: string
+}
+
+export type steamGamePlatformType = {
+  windows: boolean,
+  mac: boolean,
+  linux: boolean
+}
